feat: Added objectClasses
[Brief Changes]
Added various classes which will work with on another and
to make the code cleaner.
This is by no means done in shorter lines but it will further
improve readability.

[Object Class]
Added an Object class to store different kinds of objects.
Added various functions to support the Object class.

[Object.type Classes]
Added a Slider class with basic parameters for the Object class.
Added a Rectangle class with basic parameters for the Object class.

[Collider Class]
Added a Collider class to handle button presses in the future.