feat: added run.py
[GLOBAL]:
run.py is the ancestor of main.py you could say.
Currently I am making a lot of changes to the code
I had previously which I still take a peek at sometimes
so it's usefull to still have it as a backup to fall on
when things go wrong. Thats why it's not deleted.

[RUN.PY]
-mainloop is in place with the bare minimum
-I sorted some global and private variables
which all can still be configured ofcourse, some are just
disincouraged to be changed because it can have massive effects
on the program if not carefull.
-Pixel grid can get drawn to screen again without any problems.