feat: various additions
[General Additions]
- Color palettes
- Recent colors (WORK IN PROGRESS)

[Color Palettes]
Added color palettes so you can import your
own color palette if wanted.
This is simply done with a list with objects with different colors.
Added a collider so it acts as a button.

[Recent Colors]
A list with a maximum value which can be changed.
will work like color palettes in the future but haven't
had the time to do that.