feat: Save button
[Save Button]
You are now able to save in mode '0'
The destination file is hardcoded.

[Load Button]
Implementation is being worked on.
Import already works and you can import a JSON file into the matrix.
It will change the whole matrix to match the one of the JSON file.
I am stuck on implementing it in the frame itself, if the frame
dimensions are different then it will be difficult to change those.
Source file is hardcoded.